Getting and showing list of albums
