+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.VinylTeam8"
         tools:targetApi="31">
         <activity
-            android:name=".NavigationActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
             android:label="@string/title_activity_navigation">
             <intent-filter>
