--- app/src/main/java/com/example/vinylteam8/ui/album/AlbumViewModel.kt
@@ -1,14 +1,0 @@
-package com.example.vinylteam8.ui.album
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class AlbumViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is Album Fragment"
-    }
-    val text: LiveData<String> = _text
-}
-
