+++ app/src/main/java/com/example/vinylteam8/viewmodels/AlbumViewModel.kt
@@ -45,7 +45,7 @@ class AlbumViewModel(application: Application) :  AndroidViewModel(application) 
     }
 
     class Factory(val app: Application) : ViewModelProvider.Factory {
-        override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        override fun <T : ViewModel> create(modelClass: Class<T>): T {
             if (modelClass.isAssignableFrom(AlbumViewModel::class.java)) {
                 @Suppress("UNCHECKED_CAST")
                 return AlbumViewModel(app) as T
