+++ app/src/main/res/layout/fragment_album.xml
@@ -6,17 +6,11 @@
     android:layout_height="match_parent"
     tools:context=".ui.album.AlbumFragment">
 
-    <TextView
-        android:id="@+id/text_dashboard"
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/albumsRv"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:textAlignment="center"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:layout_height="match_parent"
+        app:layoutManager="LinearLayoutManager"
+        android:scrollbars="vertical"
+        tools:listitem="@layout/fragment_album_list"/>
 </androidx.constraintlayout.widget.ConstraintLayout>
