+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -14,7 +14,7 @@
     <fragment
         android:id="@+id/navigation_album"
         android:name="com.example.vinylteam8.ui.album.AlbumFragment"
-        android:label="@string/title_almbum"
+        android:label="@string/title_albums"
         tools:layout="@layout/fragment_album" />
 
     <fragment
